modified:   0x02-functions_nested_loops/104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -2,7 +2,7 @@
 
 int main(void)
 {
-unsigned int long first = 0, second = 1, sum = 0;
+unsigned long long first = 0, second = 1, sum = 0;
 int i;
 
 for (i = 2; i < 98; i++)
@@ -11,9 +11,9 @@ sum = first + second;
 first = second;
 second = sum;
 if (i != 98)
-printf("%ld, ", sum);
+printf("%llu, ", sum);
 else
-printf("%ld\n", sum);
+printf("%llu\n", sum);
 }
 return 0;
 }
